Fix EndPoints

+++ src/main/java/com/marketplace/product/domain/model/Product.java
@@ -30,16 +30,15 @@ public class Product {
     private String sku;
     private String name;
     private String description;
-    private Float price;
+    private Double price;
     private String imgUrl;
     private Integer unitAvailable;
-    private Float weight;
+    private Double weight;
     private String category;
     private String brand;
     private Integer amountToReserve;
     private Integer amountToCancel;
-
-
+    
     @ManyToMany//(cascade = CascadeType.ALL)
     @JoinTable(name = "PRODUCT_KEYWORD",joinColumns = {@JoinColumn(name = "PRODUCT_ID")},
             inverseJoinColumns = {@JoinColumn(name = "KEYWORD_ID")})
